backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "fastapi>=0.128.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "tencentcloud-sdk-python>=3.0.0",
    "uvicorn>=0.40.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.6",
    "psycopg[binary]>=3.3.2",
]
